still alpha version
It's still very much a beta but i'm sharing so you can see and use the math if you want. the math was the hard part.

Known issues:
floods the chat with debug info - already correcting this in the next version.
gets snagged on things pretty easily - trying to correct how I can but it's likely to always be an issue.
roll dodges sometimes, didn't think it would be this frequent.

Want to have:
Use doors better (or at all, it only worked once).
Sneak when leader sneaks.
Mount when leader mounts.
Mount when leader is over a certain distance.